Clean up motors classes.
