+++ libraries/AP_Motors/AP_MotorsBlueROV.h
@@ -1,7 +1,7 @@
 // -*- tab-width: 4; Mode: C++; c-basic-offset: 4; indent-tabs-mode: nil -*-
 
 /// @file	AP_MotorsBlueROV.h
-/// @brief	Motor control class for ROVs
+/// @brief	Motor control class for Hexacopters
 
 #ifndef __AP_MOTORS_BLUEROV_H__
 #define __AP_MOTORS_BLUEROV_H__
@@ -9,60 +9,23 @@
 #include <AP_Common/AP_Common.h>
 #include <AP_Math/AP_Math.h>        // ArduPilot Mega Vector/Matrix math Library
 #include <RC_Channel/RC_Channel.h>     // RC Channel Library
-#include "AP_MotorsMulticopter.h"
+#include "AP_Motors6DOF.h"    // Parent Motors Matrix library
 
-// pitch servo uses channel 7
-#define AP_MOTORS_CH_CAM_PITCH    CH_7
-
-/// @class      AP_MotorsTri
-class AP_MotorsBlueROV : public AP_MotorsMulticopter {
+/// @class      AP_MotorsBlueROV
+class AP_MotorsBlueROV : public AP_Motors6DOF {
 public:
 
     /// Constructor
     AP_MotorsBlueROV(uint16_t loop_rate, uint16_t speed_hz = AP_MOTORS_SPEED_DEFAULT) :
-        AP_MotorsMulticopter(loop_rate, speed_hz)
-    {
-        AP_Param::setup_object_defaults(this, var_info);
-    };
+        AP_Motors6DOF(loop_rate, speed_hz)
+    { };
 
-    // init
-    virtual void        Init();
-
-    // set update rate to motors - a value in hertz
-    void                set_update_rate( uint16_t speed_hz );
-
-    // enable - starts allowing signals to be sent to motors
-    virtual void        enable();
-
-    // output_test - spin a motor at the pwm value specified
-    //  motor_seq is the motor's sequence number from 1 to the number of motors on the frame
-    //  pwm value is an actual pwm value that will be output, normally in the range of 1000 ~ 2000
-    virtual void        output_test(uint8_t motor_seq, int16_t pwm);
-
-    // output_min - sends minimum values out to the motors
-    virtual void        output_min();
-
-    // get_motor_mask - returns a bitmask of which outputs are being used for motors or servos (1 means being used)
-    //  this can be used to ensure other pwm outputs (i.e. for servos) do not conflict
-    virtual uint16_t    get_motor_mask();
-
-    // var_info for holding Parameter information
-    static const struct AP_Param::GroupInfo var_info[];
+    // setup_motors - configures the motors for the BlueROV
+    // This takes care of all of the enabling and bitmask stuff seen in the Tricopter motors class
+    virtual void        setup_motors();
 
 protected:
-    // output - sends commands to the motors
-    void                output_armed_stabilizing();
-    void                output_armed_not_stabilizing();
-    void                output_disarmed();
 
-    // calc_yaw_radio_output - calculate final radio output for yaw channel
-    int16_t             calc_yaw_radio_output();        // calculate radio output for yaw servo, typically in range of 1100-1900
-
-    // parameters
-    AP_Int8         _yaw_servo_reverse;                 // Yaw servo signal reversing
-    AP_Int16        _yaw_servo_trim;                    // Trim or center position of yaw servo
-    AP_Int16        _yaw_servo_min;                     // Minimum angle limit of yaw servo
-    AP_Int16        _yaw_servo_max;                     // Maximum angle limit of yaw servo
 };
 
-#endif  // AP_MOTORSBLUEROV
+#endif  // AP_MotorsBlueROV
